Fixed referral migration phone number integer limitation

+++ database/migrations/2020_02_17_204215_create_referrals_table.php
@@ -16,7 +16,7 @@ class CreateReferralsTable extends Migration
         Schema::create('referrals', function (Blueprint $table) {
             $table->bigIncrements('id');
 
-            $table->integer('referee_phone')->unique();
+            $table->bigInteger('referee_phone')->unique();
             $table->string('status', 10)->default('pending');
 
             $table->timestamps();
